Crate dns_lookup

dns_lookup

A small wrapper for libc to perform simple DNS lookups.

Two main functions are provided.

lookup_host

Given a hostname, return an Iterator the IP Addresses associated with it.

use dns_lookup::lookup_host;

 let hostname = "localhost";
 let ips: Vec<std::net::IpAddr> =
   lookup_host(hostname).unwrap().collect::<std::io::Result<_>>().unwrap();
 assert!(ips.contains(&"127.0.0.1".parse().unwrap()));

lookup_addr

Given an IP Address, return the reverse DNS entry (hostname) for the given IP Address.

use dns_lookup::lookup_addr;

 let ip: std::net::IpAddr = "127.0.0.1".parse().unwrap();
 let hostname = lookup_addr(&ip).unwrap();
 assert_eq!(hostname, "localhost");

Functions

lookup_addr
Lookup the hostname of a given IP Address via DNS.

lookup_host
Lookup the address for a given hostname via DNS.