Enum dns_lookup::SockType

pub enum SockType {
    Unspec,
    Stream,
    DGram,
    Raw,
    RDM,
    SeqPacket,
    DCCP,
    Packet,
    _Other(u16),
}

Socket Type

Variants
Unspec

Unspecified (used only for hints argument in getaddrinfo).

Stream

Sequenced, reliable, connection-based byte streams.

DGram

Connectionless, unreliable datagrams of fixed max length.

Raw

Raw protocol interface.

RDM

Reliably-delivered messages.

SeqPacket

Sequenced, reliable, connection-based, datagrams of fixed maximum length.

DCCP

Datagram Congestion Control Protocol.

Packet

Linux specific way of getting packets at the dev level. For writing rarp and other similar things on the user level.

_Other(u16)

Other SockType.

It's recommended not to match on this, or create this directly, as new variants may break code depending on this variant. Instead you should use the From/Into traits to convert libc symbols directly, if the desired variant does not exist.