Module dns_lookup::lookup

Structs

LookupHost
A struct that holds a linked list of lookup results.

Enums

Error
Errors that can occur looking up a hostname.

Functions

lookup_host
Lookup a hostname via dns, return an iterator of ip addresses.