Struct dns_lookup::AddrInfo

pub struct AddrInfo {
    pub socktype: i32,
    pub protocol: i32,
    pub address: i32,
    pub sockaddr: SocketAddr,
    pub canonname: Option<String>,
    pub flags: i32,
}

Struct that stores socket information, as returned by getaddrinfo.

This maps to the same definition provided by libc backends.

Fields
socktype: i32

Type of this socket.

Values are defined by the libc on your system.

protocol: i32

Protcol family for this socket.

Values are defined by the libc on your system.

address: i32

Address family for this socket (usually matches protocol family).

Values are defined by the libc on your system.

sockaddr: SocketAddr

Socket address for this socket, usually containing an actual IP Address and port.

canonname: Option<String>

If requested, this is the canonical name for this socket/host.

flags: i32

Optional bitmask arguments, usually set to zero.