Struct dns_lookup::AddrInfoHints

pub struct AddrInfoHints {
    pub socktype: SockType,
    pub protocol: Protocol,
    pub address: AddrFamily,
    pub flags: u32,
}

A struct used as the hints argument to getaddrinfo.

Fields
socktype: SockType

Type of this socket, Unspec (0) for none.

protocol: Protocol

Protcol for this socket, IP (0) for none.

address: AddrFamily

Address family for this socket. Unspec (0) for none.

flags: u32

Optional bitmask arguments. Bitwise OR bitflags to change the behaviour of getaddrinfo. 0 for none.

The actual bitflags are not provided by this crate, and are usually exported in the libc crate. Some backends have custom flags, which may be a portability issue.